finished the foundation of the C++ client model. next is to auto-generate lwm2m objects using python and a template engine

+++ applications/lwm2m_client_cpp/src/lwm2m_object_resource.cpp
@@ -24,7 +24,7 @@ namespace nx {
     }
 
     lwm2m_object_resource::lwm2m_object_resource(uint16_t resource_id, resource_type_t res_type,
-                                                 resource_operations_t op, member_pointer mem_ptr, uint16_t mem_size) {
+                                                 resource_operations_t op, member_pointer mem_ptr, int mem_size) {
         this->resource_id = resource_id;
         this->mem_ptr = mem_ptr;
         this->member_size = mem_size;
